refactor: pkg structure
