+++ worker/txt2img.go
@@ -1,4 +1,4 @@
-package tasks
+package worker
 
 import (
 	"fmt"
